Progress on managing numi flux.  Make use of .h and .C files generated from the ntuples themselves via MakeClass to automate setting branches and creating the proper leaves.  Unfortunaely the geant3 and geant4 GNuMI ntuples have different formats.  All the most important variables are there in each, but upper/lower case varies.  A critical difference is the switch from Float_t to Double_t which means one can't simply use the same structure an change branch assignments.  The common format GNuMIFluxPassThrough needs a Copy() function defined for each type of ntuple.
The g3 case has been implemented, but the g4 version is a skeleton.
The retained variables could do with a review to ensure that we're not
carrying more than necessary.  The helper classes g3numi and g4numi
srouce (.h + .C) live in a subdirectory GNuMINtuple that needn't
otherwise be exposed to any GENIE user -- the actual code gets into
the library by having both #include "gXnumi.h" and #include "gXnumi.C".

The x-y reweight function has been integrated into GNuMIFluxPassThrough.
This has been tested against the fortran version. (this set of code
still retains some of that testing code which should be purged in
the next iteration).  Generally the values are comparable at the
expected precision (g3 32-bit float ~2.5e-5 or better in general
for the x-y weight, even better for the neutrino energy). There are
some large outliers when calculating x-y weights for muon decays
because the algorithm, as written, depends on taking the difference
in two largish numbers (e.g. ~6 - ~6 = ~0.002) which results in
a large loss of fractional error precision when starting with floats.

Some interfaces have been put in place for handling coordinate
transformations between the detector and beam, but these are not
fully implemented and/or tested.


git-svn-id: file:///home/mroda/Software/Genie/SVN/PhabGlobal/trunk@2109 cc9776de-3512-45ca-aafc-e2d9ed43c22c