Individual path segments can be further trimmed to be less than the full step (which corresponds to a volume).
The list is now internally a STL list (no longer a STL vector).  This
is in anticipation of a possible need for segment splitting if trimming
causes disjoint segments.

Both classes have provisions for cross checking ray distance and step size
values after completion of the list.  Generally I've found ray distance
errors of no more than ~8e-9 cm (80000fm) in a 42000cm swim through a
geometry.  This is on order an atomic size so one shouldn't worry about it.
This precision will allow trimming to be based on the distance travelled
along the ray.


git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/trunk@2395 cc9776de-3512-45ca-aafc-e2d9ed43c22c