Correct statement about how overhead much fetching geometry path adds (significant, not negligable)
Generalize from (lo,hi) pair to vector of (lo,hi) pairs to allow the
geometry step to be split and not just squeezed.  This might not be necessary
and how much this adds to overhead isn't quite clear (needs more testing).

New methods IsTrimmedEmpty() and GetSummedStepRange().  Also GetPosition()
to pick position within vector of (lo,hi) pairs based on fraction of total.
Needs more testing for case of split segments.


git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/trunk@2695 cc9776de-3512-45ca-aafc-e2d9ed43c22c