Implement first version of formation zone reweighting method.
This has been validatied to event generation of neutrino interactions
at ~500MeV (T2K energies) but due to small effect of FZone at these 
energies it needs further validation (comparing reweighted samples to 
regenerated samples) at higher energies as this is when FZone effects 
wil become more important.

The current scheme is only approximate for a given event sample due to the fact that 
it is impossible to calculate a weight for a hadron which has already been 
transported outside the nuclear boundary with the default formation zone 
step size. In practice this should only affect a small fraction of the 
hadrons created in the nucleus but this will depend on type of event generation
and may become more of an issue for high energy neutrinos.



git-svn-id: file:///home/mroda/Software/Genie/SVN/PhabGlobal/trunk@3187 cc9776de-3512-45ca-aafc-e2d9ed43c22c