abort with an error if LoadFluxData() fails
This commit updates gevgen_atmo to abort with an error if one of the flux files
fails to load when calling LoadFluxData(). I got bit by this where I
accidentally mistyped a single flux file and that flux failed to load, but
there was only a small warning but everything else continued to run.

In I also updated LoadFluxData() to immediately abort if a single flux
file fails to load instead of continuing in order to make sure that the error
message about the file failing to load comes last.