Massively change NHL lab-frame E calculation.
It is *very wrong* to just apply a boost to a random decay from CM
frame. Only a very small range of emission angles would actually reach
the detector (read: are valid targets for a boost to act on), meaning
the energy calculation came out VERY wrong.
Implement a dk2nu-like solution calculating the requisite boost factor
explicitly. Also acts on acceptance (same rest-to-lab boost
correction!)
Note the extra betaNHL in the factor, compared to
bsim::CalcEnuWgt. This is because of the massive NHL.