IMPORTANT FIX for timeBit. Details below.
The NHL needs to be travelling along the correct worldline, and of
course it can't reach the detector instantly! Start by guessing it
goes insanely fast and takes time (distance/c), then if the worldline
is too far away from reality, change betaLab(NHL) to be
smaller. Iterate until done.