reroll points on accCorr == 0. See more.
It is not correct to immediately discard a parent just because **one**
point is unreachable by HNL. One should attempt to redo the accCorr
procedure first.
This commit adds up to 10 rerolls. If none of these work, then we have
a bona fide unreachable point (with truly unhelpful parent momentum),
and then we bail with accCorr = 0.0