bring back the "correlated nucleon tail"
The correlated tail from previous work disappeared with the most recent modifications to the LFG model here.  Forcing the computed removal energy to be nonnegative makes sense for momenta drawn from the bulk of the LFG distribution; negative removal energy ordinarily means the nucleon is not bound (i.e. outside the nucleus).  However, for momenta drawn from the correlated tail, they will *always* be negative relative to the calculation from the LFG model: their associated KE is much larger because their *internal* (pairwise) potential energy offsets it.  Obviously the LFG computation can't be aware of that.

Here we take cases where the LFG calculation winds up negative and simply reassign them a (more) sensible removal energy.  In principle we would like something specific to SRC correlated pairs.  In the absence of that, we fall back on the mean binding energy per nucleon, which is a (bad) proxy for the one-nucleon separation energy.